+++ R/UseTemplate.R
@@ -1,3 +1,11 @@
+UseTemplateList = function(newfile, fileList, find=NULL, replace=NULL){
+for(i in fileList){
+cat(UseTemplate(i, find=find, replace=replace), file=newfile, append=TRUE)
+}
+return(invisible(NULL))
+}
+
+
 UseTemplate = function(file, find=NULL, replace=NULL){
 TXT = readLines(system.file(paste0("Templates/", file), package="BrailleR"))
 if(!is.null(replace)){